fix screenshot delete

+++ app/library/postgres.py
@@ -150,11 +150,13 @@ async def delete_asset_from_postgres(asset_id: str, session):
         )
 
         pos_key = asset_obj.screenshot_url.find("image/")
-        screenshot_key = asset_obj.screenshot_url[pos_key:]
-        s3_client.delete_object(
-            Bucket=config.get("s3.bucket_name"),
-            Key=screenshot_key,
-        )
+
+        if pos_key != -1:
+            screenshot_key = asset_obj.screenshot_url[pos_key:]
+            s3_client.delete_object(
+                Bucket=config.get("s3.bucket_name"),
+                Key=screenshot_key,
+            )
 
     delete_asset(asset_id, session)
     object_ids = get_object_ids(session)
